@@ -4,46 +4,50 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --delete-output-path",
+    "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
     "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.6",
-    "@angular/common": "^19.0.6",
-    "@angular/compiler": "^19.0.6",
-    "@angular/core": "^19.0.6",
-    "@angular/forms": "^19.0.6",
-    "@angular/platform-browser": "^19.0.6",
-    "@angular/platform-browser-dynamic": "^19.0.6",
-    "@angular/router": "^19.0.6",
-    "@ng-icons/core": "^29.5.2",
-    "@ng-icons/material-icons": "^29.5.2",
+    "@angular/animations": "^19.2.4",
+    "@angular/common": "^19.2.4",
+    "@angular/compiler": "^19.2.4",
+    "@angular/core": "^19.2.4",
+    "@angular/forms": "^19.2.4",
+    "@angular/platform-browser": "^19.2.4",
+    "@angular/platform-browser-dynamic": "^19.2.4",
+    "@angular/router": "^19.2.4",
+    "@angular/service-worker": "^19.2.4",
     "@spotify/web-api-ts-sdk": "^1.2.0",
+    "@tailwindcss/postcss": "^4.0.6",
+    "@typescript-eslint/eslint-plugin": "^8.21.0",
+    "@typescript-eslint/parser": "^8.21.0",
+    "@typescript-eslint/utils": "^8.21.0",
+    "lucide-angular": "^0.474.0",
+    "postcss": "^8.5.2",
     "rxjs": "~7.8.0",
+    "tailwindcss": "^4.0.6",
     "tslib": "^2.3.0",
+    "uuid": "^11.0.5",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.7",
-    "@angular/cli": "^19.0.7",
-    "@angular/compiler-cli": "^19.0.6",
-    "@angular/localize": "^19.0.6",
+    "@angular-devkit/build-angular": "^19.2.5",
+    "@angular/cli": "^19.2.5",
+    "@angular/compiler-cli": "^19.2.4",
+    "@angular/localize": "^19.2.4",
     "@types/jasmine": "~5.1.0",
-    "angular-eslint": "19.0.2",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^9.13.0",
+    "angular-eslint": "^19.3.0",
+    "eslint": "^9.20.1",
     "jasmine-core": "~5.1.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.14",
     "typescript": "~5.5.4",
-    "typescript-eslint": "8.10.0"
+    "typescript-eslint": "^8.10.0"
   }
 }
