@@ -1,73 +1,103 @@
-import { Component, inject } from "@angular/core";
+import { Component, inject, OnInit, signal } from "@angular/core";
 import { FormsModule } from "@angular/forms";
-import { GameStateService } from "../../services/game-state.service";
+import { StateService } from "../../services/state.service";
 import { Router, RouterLink } from "@angular/router";
-import { CardComponent } from "../../components/setup/card/card.component";
-import { CardSelectionInformation, CharacterGroup } from "../../types";
-import { SpotifyWidgetComponent } from "../../components/spotify-widget/spotify-widget.component";
+import { CharacterGroup, GameSets, GroupTypes } from "../../types";
 import { SpotifyService } from "../../services/spotify.service";
 import { environment } from "../../../environments/environment";
-import { StorageService } from "../../services/storage.service";
+import { GAME_SET_TRANSLATIONS, GROUP_TRANSLATIONS, NAME_TRANSLATIONS } from "../../i18n/translations";
+import { LucideAngularModule } from "lucide-angular";
+import { Device } from "@spotify/web-api-ts-sdk/dist/mjs/types";
+import { I18nSelectPipe } from "@angular/common";
+import { Character } from "../../models/characters/character";
 
 @Component({
     selector: "app-setup",
-    imports: [FormsModule, CardComponent, SpotifyWidgetComponent, RouterLink],
+    imports: [FormsModule, RouterLink, LucideAngularModule, I18nSelectPipe],
     templateUrl: "./setup.component.html",
     styleUrl: "./setup.component.css"
 })
-export class SetupComponent {
-    private state = inject(GameStateService);
+export class SetupComponent implements OnInit {
     private router = inject(Router);
-    private spotify = inject(SpotifyService);
-    private storage = inject(StorageService);
+    state = inject(StateService);
+    spotify = inject(SpotifyService);
 
-    public Roles: CharacterGroup[] = this.storage.SetupSelection;
-    public PeopleCount: number = this.storage.SetupPeopleCount;
+    isLoadingDevices = signal(false);
+    isSelectingDevice = signal(false);
+    isPeopleEditorOpen = signal(false);
+    isCharacterEditorOpen = signal(false);
+    grouping = signal<"group" | "game">("group");
 
-    public async StartGame(): Promise<void> {
-        this.state.Characters = this.Roles
-            .map(r => r.Cards)
-            .flat()
-            .filter(r => r.Selected)
-            .sort((ra, rb) => (ra.Character.Priority ?? -1) - (rb.Character.Priority ?? -1))
-            .map(r => r.Character);
-        this.state.People = this.state.People.slice(0, this.PeopleCount);
-        for (let i = 0; i < this.PeopleCount; i++) {
-            const person = this.state.People[i];
-            this.state.People[i] = {
-                Character: undefined,
-                IsVictim: false,
-                IsProtected: false,
-                IsEnchanted: false,
-                IsDead: false,
-                IsWerewolf: false,
-                Id: i,
-                Name: person?.Name ?? ""
-            };
+    NAME_TRANSLATIONS = NAME_TRANSLATIONS;
+
+    ngOnInit(): void {
+        if (this.state.People.length <= 0) {
+            for (let i = 0; i < 8; i++) {
+                this.state.addPerson();
+            }
         }
 
-        this.state.StartGame();
-        if (this.spotify.IsAuthenticated && this.spotify.CurrentDevice && !this.spotify.BackgroundMusicStarted) {
+        if (this.state.SelectedCharacters.length <= 0) {
+            this.state.SelectedCharacters = this.state.AllCharacters.filter(c => c.Game === GameSets.BaseGame);
+        }
+
+        this.state.InGame = false;
+    }
+
+    async updateDeviceList(): Promise<void> {
+        this.isLoadingDevices.set(true);
+        await this.spotify.UpdateDeviceList().then(() => {
+            setTimeout(() => {
+                this.isLoadingDevices.set(false);
+            }, 1000);
+        });
+    }
+
+    async setDevice(device: Device | undefined): Promise<void> {
+        this.isSelectingDevice.set(true);
+        await this.spotify.SetDevice(device?.id ?? undefined, false).then(() => {
+            this.isSelectingDevice.set(false);
+        });
+    }
+
+    public async startBackgroundMusic(): Promise<void> {
+        await this.spotify.PlayPlaylist(environment.spotify.playlists.start, false);
+        this.state.MusicStarted = true;
+        await this.spotify.UpdatePlaybackState();
+    }
+
+    public getCharacterGroup(): CharacterGroup[] {
+        if (this.grouping() === "group") {
+            return Object.values(GroupTypes).filter(v => typeof v === "number").map(type => {
+                return {
+                    Name: GROUP_TRANSLATIONS[type],
+                    Cards: this.state.AllCharacters.filter((c: Character) => c.Group === type)
+                }
+            });
+        }
+        return Object.values(GameSets).filter(v => typeof v === "number").map(gameSet => {
+            return {
+                Name: GAME_SET_TRANSLATIONS[gameSet],
+                Cards: this.state.AllCharacters.filter((c: Character) => c.Game === gameSet)
+            }
+        });
+    }
+
+    toggleCharacterSelection(character: Character): void {
+        if (this.state.SelectedCharacters.includes(character)) {
+            this.state.SelectedCharacters = this.state.SelectedCharacters.filter(c => c !== character);
+        } else {
+            this.state.SelectedCharacters = [...this.state.SelectedCharacters, character];
+        }
+    }
+
+    async startGame() {
+        this.state.SelectedCharacters = this.state.SelectedCharacters;
+        this.state.startGame();
+        if (this.spotify.IsAuthenticated && this.spotify.CurrentDevice && !this.state.MusicStarted) {
             await this.spotify.PlayPlaylist(environment.spotify.playlists.start, false);
-            this.spotify.BackgroundMusicStarted = true;
+            this.state.MusicStarted = true;
         }
         this.router.navigateByUrl("/narrator");
-    }
-
-    public HandleSelectionStateChange(card: CardSelectionInformation, selected: boolean): void {
-        card.Selected = selected;
-        this.storage.SetupSelection = this.Roles;
-    }
-
-    public HandlePeopleCountChanged(people: number): void {
-        this.PeopleCount = people;
-        this.storage.SetupPeopleCount = people;
-    }
-
-    public GetSelectedCards(): CardSelectionInformation[] {
-        return this.Roles
-            .map(r => r.Cards)
-            .flat()
-            .filter(r => r.Selected);
     }
 }
