@@ -1,9 +1,10 @@
 import { Component } from "@angular/core";
 import { RouterLink } from "@angular/router";
+import { LucideAngularModule } from "lucide-angular";
 
 @Component({
     selector: "app-success",
-    imports: [RouterLink],
+    imports: [RouterLink, LucideAngularModule],
     templateUrl: "./success.component.html"
 })
 export class SuccessComponent {
