@@ -1,6 +1,7 @@
-import { Injectable } from "@angular/core";
+import { inject, Injectable } from "@angular/core";
 import { environment } from "../../environments/environment";
 import { DefaultResponseDeserializer, Device, PlaybackState, SpotifyApi } from "@spotify/web-api-ts-sdk";
+import { StateService } from "./state.service";
 
 const BROKEN_URLS = [
     "me/player/shuffle",
@@ -16,6 +17,7 @@ const BROKEN_URLS = [
 })
 export class SpotifyService {
     private sdk: SpotifyApi;
+    private state = inject(StateService);
 
     private clientId = environment.spotify.clientId;
     private redirectUri = environment.spotify.redirectUri;
@@ -24,7 +26,6 @@ export class SpotifyService {
     public AvailableDevices: Device[] = [];
     public CurrentDevice: Device | null = null;
     public PlaybackState: PlaybackState | null = null;
-    public BackgroundMusicStarted = false;
 
     constructor() {
         this.sdk = SpotifyApi.withUserAuthorization(this.clientId, this.redirectUri, [
@@ -43,20 +44,36 @@ export class SpotifyService {
                 }
             },
         });
+
         this.sdk.getAccessToken().then((token) => {
             this.IsAuthenticated = Boolean(token);
+            if (this.IsAuthenticated) {
+                this.startPlaybackUpdater();
 
-            this.UpdateDeviceList();
+                this.UpdateDeviceList();
+                if (this.state.CurrentDeviceId) {
+                    this.UpdateDeviceList().then(() => {
+                        this.SetDevice(this.state.CurrentDeviceId, undefined);
+                    });
+                }
+            }
+
         });
+    }
 
-        setInterval(async () => {
-            console.table({ authenticated: this.IsAuthenticated, currentDevice: Boolean(this.CurrentDevice), backgroundMusicStarted: this.BackgroundMusicStarted, playbackState: { Values: this.PlaybackState } });
-            if (this.IsAuthenticated && this.CurrentDevice && this.BackgroundMusicStarted) {
+    private startPlaybackUpdater() {
+        const int = setInterval(async () => {
+            if (!this.IsAuthenticated || !this.CurrentDevice) {
+                clearInterval(int);
+                return;
+            }
+
+            if (this.IsAuthenticated && this.CurrentDevice && this.state.MusicStarted) {
                 await this.UpdatePlaybackState();
                 if (this.PlaybackState?.progress_ms && this.PlaybackState?.item?.duration_ms) {
                     const remainingTime = this.PlaybackState.item.duration_ms - this.PlaybackState.progress_ms;
 
-                    if ((remainingTime + 500) < 10_000) {
+                    if ((remainingTime + 500) < 10000) {
                         setTimeout(async () => {
                             await this.UpdatePlaybackState();
                         }, remainingTime + 500);
@@ -75,7 +92,10 @@ export class SpotifyService {
             const profile = await this.sdk.currentUser.profile();
             this.IsAuthenticated = Boolean(profile);
 
-            this.UpdateDeviceList();
+            if (this.IsAuthenticated) {
+                this.startPlaybackUpdater();
+                this.UpdateDeviceList();
+            }
         } catch {
             this.IsAuthenticated = false;
         }
@@ -88,7 +108,7 @@ export class SpotifyService {
     }
 
     public async UpdateDeviceList() {
-        if (!this.IsAuthenticated) {
+        if (!this.checkAuthentication()) {
             return;
         }
 
@@ -100,19 +120,20 @@ export class SpotifyService {
         }
     }
 
-    public async SetDevice(device: Device) {
-        if (!this.IsAuthenticated || !device.id) {
+    public async SetDevice(deviceId: string | undefined, play: boolean | undefined) {
+        if (!this.checkCurrentDevice(deviceId) || !this.checkAuthentication()) {
             return;
         }
 
-        await this.sdk.player.transferPlayback([device.id], false);
-        this.CurrentDevice = device;
+        await this.sdk.player.transferPlayback([deviceId], play);
+        this.CurrentDevice = this.AvailableDevices.find(device => device.id === deviceId) ?? null;
+        this.state.CurrentDeviceId = deviceId;
 
         await this.UpdatePlaybackState();
     }
 
     public async UpdatePlaybackState() {
-        if (!this.IsAuthenticated) {
+        if (!this.checkAuthentication()) {
             return;
         }
 
@@ -125,7 +146,7 @@ export class SpotifyService {
     }
 
     public async PlayPlaylist(playlist: string, shuffle = true) {
-        if (!this.IsAuthenticated || !this.CurrentDevice) {
+        if (!this.checkCurrentDevice(this.CurrentDevice?.id) || !this.checkAuthentication()) {
             return;
         }
 
@@ -136,7 +157,7 @@ export class SpotifyService {
     }
 
     public async Pause() {
-        if (!this.IsAuthenticated || !this.CurrentDevice) {
+        if (!this.checkCurrentDevice(this.CurrentDevice?.id) || !this.checkAuthentication()) {
             return;
         }
 
@@ -144,7 +165,7 @@ export class SpotifyService {
     }
 
     public async Resume() {
-        if (!this.IsAuthenticated || !this.CurrentDevice) {
+        if (!this.checkCurrentDevice(this.CurrentDevice?.id) || !this.checkAuthentication()) {
             return;
         }
 
@@ -152,7 +173,7 @@ export class SpotifyService {
     }
 
     public async SkipSong() {
-        if (!this.IsAuthenticated || !this.CurrentDevice) {
+        if (!this.checkCurrentDevice(this.CurrentDevice?.id) || !this.checkAuthentication()) {
             return;
         }
 
@@ -160,12 +181,43 @@ export class SpotifyService {
     }
 
     public async QueueSongRandom(playlist: string) {
-        if (!this.IsAuthenticated || !this.CurrentDevice) {
+        if (!this.checkCurrentDevice(this.CurrentDevice?.id) || !this.checkAuthentication()) {
             return;
         }
 
         const tracks = (await this.sdk.playlists.getPlaylistItems(playlist)).items;
         const track = tracks[Math.floor(Math.random() * tracks.length)].track;
         await this.sdk.player.addItemToPlaybackQueue(track.uri, this.CurrentDevice.id!);
+    }
+
+    public getIconOfDevice(device: Device): string {
+        switch (device.type.toLowerCase()) {
+            case "computer": return "laptop";
+            case "smartphone": return "smartphone";
+            case "tablet": return "tablet";
+            case "tv": return "tv";
+            case "automobile": return "car-front";
+            default: return "speaker";
+        }
+    }
+
+    private checkAuthentication() {
+        if (!this.IsAuthenticated) {
+            this.CurrentDevice = null;
+            this.state.CurrentDeviceId = undefined;
+            this.state.MusicStarted = false;
+            return false;
+        }
+        return true;
+    }
+
+    private checkCurrentDevice(currentDeviceId: string | null | undefined): currentDeviceId is string {
+        if (!currentDeviceId) {
+            this.CurrentDevice = null;
+            this.state.CurrentDeviceId = undefined;
+            this.state.MusicStarted = false;
+            return false;
+        }
+        return true;
     }
 }
