@@ -1,6 +1,5 @@
 import { Injectable, isDevMode } from "@angular/core";
 import { CharacterGroup } from "../types";
-import { GroupedRoles } from "../models/characters";
 
 @Injectable({
     providedIn: "root"
@@ -9,19 +8,20 @@ export class StorageService {
     // Setup
     public get SetupSelection(): CharacterGroup[] {
         const selection: boolean[][] = JSON.parse(localStorage.getItem("setup_selected_cards") ?? "[]");
-        const selectionGroups = GroupedRoles;
-        for (let group = 0; group < selectionGroups.length; group++) {
-            for (let selectionIdx = 0; selectionIdx < selectionGroups[group].Cards.length; selectionIdx++) {
-                if (selection?.[group]?.[selectionIdx]) {
-                    selectionGroups[group].Cards[selectionIdx].Selected = selection[group][selectionIdx];
-                }
-            }
-        }
-        return selectionGroups;
+        // const selectionGroups = GroupedRoles;
+        // for (let group = 0; group < selectionGroups.length; group++) {
+        //     for (let selectionIdx = 0; selectionIdx < selectionGroups[group].Cards.length; selectionIdx++) {
+        //         if (selection?.[group]?.[selectionIdx]) {
+        //             selectionGroups[group].Cards[selectionIdx].Selected = selection[group][selectionIdx];
+        //         }
+        //     }
+        // }
+        // return selectionGroups;
+        return [];
     }
     public set SetupSelection(selectionGroups: CharacterGroup[]) {
-        const slection = selectionGroups.map(v => v.Cards.map(c => c.Selected));
-        localStorage.setItem("setup_selected_cards", JSON.stringify(slection));
+        // const slection = selectionGroups.map(v => v.Cards.map(c => c.Selected));
+        // localStorage.setItem("setup_selected_cards", JSON.stringify(slection));
     }
 
     public get SetupPeopleCount(): number {
