@@ -1,22 +1,10 @@
+import { Character } from "./models/characters/character";
+import { Person as Person } from "./models/state/person";
 import { DialogService } from "./services/dialog.service";
-import { GameStateService } from "./services/game-state.service";
+import { StateService } from "./services/state.service";
 
-export interface Person {
-    Id: number;
-    Name: string | undefined;
-    IsProtected: boolean;
-    IsVictim: boolean;
-    IsEnchanted: boolean;
-    IsWerewolf: boolean;
-    IsDead: boolean;
-    Character?: Character;
-}
-
-export interface Connection {
-    Type: ConnectionTypes;
-    From: Person;
-    To: Person;
-}
+export type Connection = { ConnectionType: ConnectionTypes, From: Person, To: Person };
+export type Connections = Map<Partial<ConnectionTypes>, Connection>;
 
 export enum ConnectionTypes {
     Love,
@@ -24,11 +12,10 @@ export enum ConnectionTypes {
     Sleepover
 }
 
-export interface Action {
-    Name: string;
-    Image: string;
-    GetPoints?: () => (string | false)[]
-    GetButtons?: () => ActionButton[]
+export interface ActionProvider {
+    Id: string,
+    GetActions: () => (string | false)[]
+    GetButtons: () => ActionButton[]
 }
 
 export interface ActionButton {
@@ -37,26 +24,120 @@ export interface ActionButton {
 }
 
 export type ActionCallback = (services: {
-    GameState: GameStateService,
+    GameState: StateService,
     Dialog: DialogService,
 }) => void;
 
-export interface Character {
-    Name: string;
-    Image: string;
-    IsSingle: boolean;
-    Priority?: number;
-    Action?: Action;
-    AssignedPerson?: Person
-    AssignedPeople?: Person[]
-    IsAwakeThisNight: (nightCount: number, gameState: GameStateService) => boolean;
+export enum GroupTypes {
+    Wolves,
+    Active,
+    Passive,
+    Loners
+}
+
+export enum GameSets {
+    BaseGame,
+    NewMoon,
+    Characters,
+    Special
 }
 
 export interface CharacterGroup {
     Name: string;
-    Cards: CardSelectionInformation[]
+    Cards: Character[]
 }
 
-export interface CardSelectionInformation {
-    Character: Character, Selected: boolean
+export enum BasePriority {
+    Initial = 0,
+    Wolf = 50,
+    PostWolf = 100
+}
+
+export enum ChangeReason {
+    DiedGotKilled,
+    DiedSleptWithVictim,
+    DiedOfBrokenHeart,
+    TurnedIntoWolf,
+}
+
+export interface DaybreakChange {
+    person: Person,
+    reason: ChangeReason,
+    apply: () => void
+}
+
+export interface AppState {
+    inGame: boolean,
+    spotify: {
+        currentDevice: string | undefined,
+        musicStarted: boolean,
+    }
+}
+
+export interface State {
+    app: {
+        inGame: boolean,
+    },
+    game: {
+        round: number,
+        people: Person[],
+        selectedCharacters: Character[],
+        connections: Connections
+    },
+    music: {
+        currentDevice: string | undefined,
+        musicStarted: boolean,
+    }
+}
+
+export interface Point {
+    x: number,
+    y: number
+}
+
+export enum DialogTypes {
+    PeopleSelection,
+    PersonDetails,
+    CharacterDetails,
+    Confirm
+}
+
+export interface PeopleSelectionDialogData {
+    title: string,
+    numberOfPeople: number | undefined,
+    people: Person[]
+}
+
+export interface PersonDetailsDialogData {
+    person: Person
+}
+
+export interface CharacterDetailsDialogData {
+    character: Character
+}
+
+export interface ConfirmDialogData {
+    title: string,
+}
+
+export type DialogData = {
+    type: DialogTypes.PeopleSelection,
+    data: PeopleSelectionDialogData,
+    res: (people: Person[]) => void,
+    rej: () => void
+} | {
+    type: DialogTypes.PersonDetails,
+    data: PersonDetailsDialogData,
+    res: () => void,
+    rej: () => void
+} | {
+    type: DialogTypes.CharacterDetails,
+    data: CharacterDetailsDialogData,
+    res: () => void,
+    rej: () => void
+} | {
+    type: DialogTypes.Confirm,
+    data: ConfirmDialogData,
+    res: (response: boolean) => void,
+    rej: () => void
 }
